requirements.txt
numpy==1.20.3
Pillow==9.5.0
PyYAML==6.0
scikit_learn==1.2.2
scipy==1.7.1
torch==1.12.1
torchvision==0.15.1
